Made changes suggested by Monica L. Henry.
Under Application Access – Change “On Box…” to “In Box… ”
o   They may not know what “Group membership” means. I wonder if we should call this “eDoc workflow membership”
o   I love that you moved the eApp and eDoc groups, as well as the aging reports to Box. Is it possible for that section for you to open a new window if they click on it? As it is you have to back arrow several times to get back to the form.
·         Under Admissions Application
o   Training is spelled wrong
·         Awards and Fellowships
o   What does “Fellowship Contact Only” meaning?
·         Bulletin
o   Add “for IU Graduate School Programs only” under Bulletin Contact
·         Graduate School Student Progress eDoc Approver section – can we add IU to the title? We, the IUPUI Graduate Office, also handle Purdue eDoc access.
o   Add “Master’s” before “Application and Recommendation…”
o   I’m not sure that Secretary Recorder and DGS should be in this section
·         IUPUI Graduate Office Roles sectnno
o   Does IUB not need this information? We utilize the same contact database?